Fix: 2021-09-21

+++ xabstractdebugger.cpp
@@ -663,24 +663,23 @@ QMap<QString, QVariant> XAbstractDebugger::getRegisters(void *hThread, REG_OPTIO
         mapResult.insert("EIP",(quint32)(context.Eip));
         mapResult.insert("ESP",(quint32)(context.Esp));
     #else
-        mapResult.insert("RDI",(quint32)(context.Rdi));
-        mapResult.insert("RSI",(quint32)(context.Rsi));
-        mapResult.insert("RBX",(quint32)(context.Rbx));
-        mapResult.insert("RDX",(quint32)(context.Rdx));
-        mapResult.insert("RCX",(quint32)(context.Rcx));
-        mapResult.insert("RAX",(quint32)(context.Rax));
-        mapResult.insert("RBP",(quint32)(context.Rbp));
-        mapResult.insert("RIP",(quint32)(context.Rip));
-        mapResult.insert("RSP",(quint32)(context.Rsp));
-
-        mapResult.insert("R8",(quint32)(context.R8));
-        mapResult.insert("R9",(quint32)(context.R9));
-        mapResult.insert("R10",(quint32)(context.R10));
-        mapResult.insert("R11",(quint32)(context.R11));
-        mapResult.insert("R12",(quint32)(context.R12));
-        mapResult.insert("R13",(quint32)(context.R13));
-        mapResult.insert("R14",(quint32)(context.R14));
-        mapResult.insert("R15",(quint32)(context.R15));
+        mapResult.insert("RDI",(quint64)(context.Rdi));
+        mapResult.insert("RSI",(quint64)(context.Rsi));
+        mapResult.insert("RBX",(quint64)(context.Rbx));
+        mapResult.insert("RDX",(quint64)(context.Rdx));
+        mapResult.insert("RCX",(quint64)(context.Rcx));
+        mapResult.insert("RAX",(quint64)(context.Rax));
+        mapResult.insert("RBP",(quint64)(context.Rbp));
+        mapResult.insert("RIP",(quint64)(context.Rip));
+        mapResult.insert("RSP",(quint64)(context.Rsp));
+        mapResult.insert("R8",(quint64)(context.R8));
+        mapResult.insert("R9",(quint64)(context.R9));
+        mapResult.insert("R10",(quint64)(context.R10));
+        mapResult.insert("R11",(quint64)(context.R11));
+        mapResult.insert("R12",(quint64)(context.R12));
+        mapResult.insert("R13",(quint64)(context.R13));
+        mapResult.insert("R14",(quint64)(context.R14));
+        mapResult.insert("R15",(quint64)(context.R15));
     #endif
         mapResult.insert("EFLAGS",(quint32)(context.EFlags));
     }
